+++ src/main/java/ar/com/leo/super_master_backend/dominio/catalogo/service/CatalogoServiceImpl.java
@@ -6,8 +6,10 @@ import ar.com.leo.super_master_backend.dominio.catalogo.dto.CatalogoUpdateDTO;
 import ar.com.leo.super_master_backend.dominio.catalogo.entity.Catalogo;
 import ar.com.leo.super_master_backend.dominio.catalogo.mapper.CatalogoMapper;
 import ar.com.leo.super_master_backend.dominio.catalogo.repository.CatalogoRepository;
+import ar.com.leo.super_master_backend.dominio.common.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -19,6 +21,7 @@ public class CatalogoServiceImpl implements CatalogoService {
     private final CatalogoMapper mapper;
 
     @Override
+    @Transactional(readOnly = true)
     public List<CatalogoDTO> listar() {
         return repo.findAll()
                 .stream()
@@ -27,13 +30,15 @@ public class CatalogoServiceImpl implements CatalogoService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public CatalogoDTO obtener(Integer id) {
         return repo.findById(id)
                 .map(mapper::toDTO)
-                .orElseThrow(() -> new RuntimeException("Catálogo no encontrado"));
+                .orElseThrow(() -> new NotFoundException("Catálogo no encontrado"));
     }
 
     @Override
+    @Transactional
     public CatalogoDTO crear(CatalogoCreateDTO dto) {
         Catalogo entity = mapper.toEntity(dto);
         repo.save(entity);
@@ -41,10 +46,11 @@ public class CatalogoServiceImpl implements CatalogoService {
     }
 
     @Override
+    @Transactional
     public CatalogoDTO actualizar(Integer id, CatalogoUpdateDTO dto) {
 
         Catalogo entity = repo.findById(id)
-                .orElseThrow(() -> new RuntimeException("Catálogo no encontrado"));
+                .orElseThrow(() -> new NotFoundException("Catálogo no encontrado"));
 
         mapper.updateEntityFromDTO(dto, entity);
 
@@ -54,7 +60,11 @@ public class CatalogoServiceImpl implements CatalogoService {
     }
 
     @Override
+    @Transactional
     public void eliminar(Integer id) {
+        if (!repo.existsById(id)) {
+            throw new NotFoundException("Catálogo no encontrado");
+        }
         repo.deleteById(id);
     }
